[api] Refactor ing to snake_case attrs and others
* JSON: data type "directory" to "dir"
* remove DO.testing in attrs
* Rename datatype.py to data_type.py
