+++ python/rain/common/data_type.py
@@ -1,0 +1,7 @@
+import enum
+
+
+class DataType(enum.Enum):
+
+    BLOB = "blob"
+    DIRECTORY = "dir"
